Day 4 Fin

+++ app/Http/Controllers/ProductController.php
@@ -26,17 +26,31 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => 'required|min:5',
-            'slug' => 'required',
-            'price' => 'required',
-        ]);
-        return Product::create($request->all());
+        // Role
+        $user = auth()->user();
+
+        if ($user->tokenCan("1"))
+        {
+            $request->validate([
+                'name'  => 'required|min:5',
+                'slug'  => 'required',
+                'price' => 'required',
+            ]);
+
+            return Product::create($request->all());
+        }
+        else
+        {
+            return [
+                "status" => 'Permission denied to create',
+
+            ];
+        }
     }
 
     /**
      * Display the specified resource.
-     * 
+     *
      * @param  int $id
      * @return \Illuminate\Http\Response
      */
@@ -54,10 +68,24 @@ class ProductController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $product = Product::find($id);
-        $product->update($request->all());
+        $user = auth()->user();
 
-        return $product;
+        if ($user->tokenCan("1"))
+        {
+
+            $product = Product::find($id);
+            $product->update($request->all());
+
+            return $product;
+
+        }
+        else
+        {
+            return [
+                "status" => 'Permission denied to create',
+
+            ];
+        }
     }
 
     /**
@@ -68,6 +96,19 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
-        return Product::destroy($id);
+        $user = auth()->user();
+
+        if ($user->tokenCan("1"))
+        {
+
+            return Product::destroy($id);
+        }
+        else
+        {
+            return [
+                "status" => 'Permission denied to create',
+
+            ];
+        }
     }
 }
